+++ NEWS.md
@@ -1,5 +1,12 @@
 # log history of echarty package development
 
+## v. 0.2.0
+
+- removed dependencies for Shiny and crosstalk
+- js2r and eshiny are now R demos
+- change leaflet plugin source.data to dataset, preset center and zoom
+- built with latest R v.4.1.0
+
 ## v. 0.1.4
 
 - upgrade to ECharts v.5.1.1